Fix(BigInteger) : Goods 데이터 타입 변경 (int > BigInteger) - #19
2c 까지만 저장되던 이슈 해결

+++ Assets/IslandMonkey/Scripts/UI/GoodsViewModel.cs
@@ -1,10 +1,12 @@
+using System.Numerics;
+
 namespace IslandMonkey.MVVM
 {
 	public class GoodsViewModel : ViewModel
 	{
-		public int Gold { get; set; }
-		public int Banana { get; set; }
-		public int Clam { get; set; }
+		public BigInteger Gold { get; set; }
+		public BigInteger Banana { get; set; }
+		public BigInteger Clam { get; set; }
 
 		protected override void Start()
 		{
